Update REACT_BASICS.md
Added "import "./Kennel.css"" to the child component instructions. It was the only line excluded in the updated instructions and could be easily removed if the student pastes over all their previous code with the updated example.

Also removed a line space in that component to be more consistent with earlier examples formatting.