Removed small typos from first chapters in React unit.
In REACT_BASICS.md: Added "import "./Kennel.css"" to the child component instructions. It was the only line excluded in the updated instructions and could be easily removed if the student pastes over all their previous code with the updated example.

Removed a line space in that component to be more consistent with earlier examples formatting.

Changed names of location, customer, and employee component JS files in final practice to match the animal component example file names earlier.

In PROPS_STATE_INTRO.md: Removed errant "," and moved the word "to" outside of the syntax highlighting for the function name.